Rename package compression to compress
