+++ clients/src/main/java/org/apache/kafka/common/compress/KafkaLZ4BlockOutputStream.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.kafka.common.compression;
+package org.apache.kafka.common.compress;
 
 import java.io.IOException;
 import java.io.OutputStream;
