+++ clients/src/main/java/org/apache/kafka/common/compress/SnappyFactory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.common.compression;
+package org.apache.kafka.common.compress;
 
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.utils.ByteBufferInputStream;
