+++ clients/src/main/java/org/apache/kafka/common/compress/ZstdFactory.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.common.compression;
+package org.apache.kafka.common.compress;
 
 import com.github.luben.zstd.RecyclingBufferPool;
 import com.github.luben.zstd.ZstdInputStreamNoFinalizer;
