+++ clients/src/main/java/org/apache/kafka/common/record/CompressionType.java
@@ -17,10 +17,10 @@
 package org.apache.kafka.common.record;
 
 import org.apache.kafka.common.KafkaException;
-import org.apache.kafka.common.compression.KafkaLZ4BlockInputStream;
-import org.apache.kafka.common.compression.KafkaLZ4BlockOutputStream;
-import org.apache.kafka.common.compression.SnappyFactory;
-import org.apache.kafka.common.compression.ZstdFactory;
+import org.apache.kafka.common.compress.KafkaLZ4BlockInputStream;
+import org.apache.kafka.common.compress.KafkaLZ4BlockOutputStream;
+import org.apache.kafka.common.compress.SnappyFactory;
+import org.apache.kafka.common.compress.ZstdFactory;
 import org.apache.kafka.common.utils.ByteBufferInputStream;
 import org.apache.kafka.common.utils.ByteBufferOutputStream;
